Factor common mongodb setup out from test cases
Also fix flake8 errors
