+++ tests/test_sample_data.py
@@ -3,33 +3,20 @@
 This test suite requires a local mongodb instance and writes to/drops a
 database named "abe-unittest" for testing.
 """
-import os
 import unittest
-from pymongo import MongoClient
-import logging
-from .context import abe
+
 from abe import sample_data
-import pdb
+
+from . import abe_unittest
+from .context import abe  # noqa: F401
 
 
-class SampleDataTestCase(unittest.TestCase):
-
-    def setUp(self):
-        """Setup testing environment"""
-        os.environ["DB_NAME"] = "abe-unittest"
-        os.environ["MONGO_URI"] = ""
-        from abe import database as db
-        self.db = db
-
-    def tearDown(self):
-        """Teardown testing environment"""
-        client = MongoClient()
-        client.drop_database(os.environ['DB_NAME'])  # remove testing db
-        client.close()
+class SampleDataTestCase(abe_unittest.TestCase):
 
     def test_load_data(self):
         """Load sample data"""
         sample_data.load_data(self.db)
+
 
 if __name__ == '__main__':
     unittest.main()
